+++ src/cli/actions/object_get.rs
@@ -5,7 +5,7 @@ use crate::{
 use anyhow::{anyhow, Context, Result};
 use colored::Colorize;
 use std::{cmp::min, path::Path};
-use tokio::{fs::File, io::AsyncWriteExt};
+use tokio::{fs::OpenOptions, io::AsyncWriteExt};
 
 pub async fn handle(s3: &S3, action: Action) -> Result<()> {
     if let Action::GetObject {
@@ -13,6 +13,7 @@ pub async fn handle(s3: &S3, action: Action) -> Result<()> {
         get_head,
         dest,
         quiet,
+        force,
     } = action
     {
         if get_head {
@@ -39,21 +40,33 @@ pub async fn handle(s3: &S3, action: Action) -> Result<()> {
                 |d| Path::new(&d).join(file_name),
             );
 
-            // TODO implement  -f --force
-            if path.is_file() {
+            // check if file exists
+            if path.is_file() && !force {
                 return Err(anyhow!("file {:?} already exists", path));
+            }
+
+            // open
+            let mut options = OpenOptions::new();
+            options.write(true).create(true);
+
+            // Set truncate flag to overwrite the file if it exists
+            if force {
+                options.truncate(true);
             }
 
             // do the request
             let action = actions::GetObject::new(&key);
             let mut res = action.request(s3).await?;
-            let mut file = File::create(&path).await?;
+
+            // Open the file with the specified options
+            let mut file = options.open(path).await?;
 
             // get the file_size in bytes by using the content_length
             let file_size = res
                 .content_length()
                 .context("could not get content_length")?;
 
+            // if quiet is true, then use a default progress bar
             let pb = if quiet {
                 Bar::default()
             } else {
@@ -63,10 +76,13 @@ pub async fn handle(s3: &S3, action: Action) -> Result<()> {
             let mut downloaded = 0;
             while let Some(bytes) = res.chunk().await? {
                 let new = min(downloaded + bytes.len() as u64, file_size);
+
                 downloaded = new;
+
                 if let Some(pb) = pb.progress.as_ref() {
                     pb.set_position(new);
                 }
+
                 file.write_all(&bytes).await?;
             }