+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.7.1
 * Removed after_help
+* Added `-f/--force` when getting a file
 
 ## 0.7.0 🚜
 * Added support for Additional Checksum Algorithms `x-amz-checksum-mode` option `--checksum`